Add a toolbar & menu for about app and share champions URL
Included layouts in the XML layouts should be the last items inside the
tag, otherwise it will thrown an error.

Getting the ID of the included layouts with View Binding also should
passed with the source ID. For example, to get the toolbar that included with ID
`toolbar` should be `binding.<idOfIncludedLayout>.<idOfincludedLayoutSource>`
like `binding.toolbarDetail.toolbar`.

All available colors, assets, and attributes value built-in from Android can be
called with `@android:` and `?attr/` in XML layouts.

References:
- https://developer.android.com/develop/ui/views/components/appbar
- https://developer.android.com/guide/topics/resources/menu-resource
- https://developer.android.com/training/sharing
- https://developer.android.com/reference/androidx/appcompat/widget/ShareActionProvider
- https://guides.codepath.com/android/Sharing-Content-with-Intents
- https://guides.codepath.com/android/Defining-The-ActionBar
- https://guides.codepath.com/android/Using-the-App-Toolbar
- https://github.com/codepath/android_guides/wiki/Using-the-App-ToolBar
- https://www.vogella.com/tutorials/AndroidActionBar/article.html